requirements.txt
# Core Detection & Tracking
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
supervision>=0.16.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Dashboard
streamlit>=1.28.0

# Geospatial & Analytics
shapely>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
filterpy>=1.4.5
folium>=0.15.0
pyproj>=3.6.0
rasterio>=1.3.0
streamlit-folium>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
imageio>=2.31.0

# Reporting
reportlab>=4.0.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0.0
